Compile under Linux
git-svn-id: svn://localhost/steel.new/trunk@254 99e027b2-380c-0410-a737-a5784e1a7bde

+++ steel/trunk/code/graph/opengl/opengl_glsl.cpp
@@ -128,7 +128,7 @@ void GLSL::loadLog(GLuint object)
     if(logLength < 1 )
         return;
                                     // try to avoid allocating buffer
-    if ( logLength > sizeof ( buffer ) )
+    if ( logLength > (int)sizeof ( buffer ) )
     {
         infoLog = (GLcharARB*) malloc ( logLength );
